+++ src/modules/discord_bot.py
@@ -22,23 +22,6 @@ from src.modules.utils import GroupCreator
 
 DB = Firebase()
 
-
-def authorization_required(func):
-    @wraps(func)
-    async def wrapper(*args, **kwargs):
-        import src.texts.auth as txt
-        ctx = args[1]
-        user = DB.get_user(discord_id=ctx.message.author.id)
-        if user is None:
-            logging.info("Usuario no registrado")
-            await ctx.send(txt.NOT_REGISTERED_ERROR)
-            return
-        logging.info(f"Usuario registrado")
-        return await func(*args)
-
-    return wrapper
-
-
 def group_required(func):
     @wraps(func)
     async def wrapper(*args, **kwargs):
